requirements.txt
torch
pandas
numpy
tqdm
pyyaml
matplotlib
wget
linear_attention_transformer
h5py
